build: finonacci to n50

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,23 +9,23 @@
 
 int main(void)
 {
-	unsigned int count, term1, term2, fibn;
+	long long unsigned int count, term1, term2, fibn;
 
 	term1 = 1;
 	term2 = 2;
-	printf("%d, %d, ", term1, term2);
+	printf("%llu, %llu, ", term1, term2);
 
 	for (count = 2; count < 50; count++)
 	{
 		if (count < 49)
 		{
 			fibn = term1 + term2;
-			printf("%u, ", fibn);
+			printf("%llu, ", fibn);
 			term1 = term2;
 			term2 = fibn;
 		}
 	}
 	fibn = term1 + term2;
-	printf("%u\n", fibn);
+	printf("%lu\n", fibn);
 	return (0);
 }
